eerly_studio/src/eerly_studio/insights/pyproject.toml
[project]
name = "insights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b-code-interpreter>=2.4.1",
    "fastparquet>=2025.12.0",
    "langchain>=1.2.10",
    "langchain-core>=0.3.0",
    "langchain-openai>=1.1.10",
    "langfuse>=3.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=1.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "psycopg[binary,pool]>=3.2.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
    "sqlglot>=29.0.1",
    "tenacity>=9.1.4",
]

[tool.uv]
package = false
